Added the this. to TheText setter
will all the current libraries and frameworks, with out the `this` in front of `RaiseAndSetIfChanged(ref theText, value):` I had a compiler error that my TheViewModel did not have matching signature.
I noticed all the other examples included the `this.` so I added it here and it worked.

To be honest I'm not totally sure why the `this` is required. In fact it kinda bugs me that it is required to compile.  As I understand inheritance the `this` shouldn't be required.