add a definition for OAPH before it is first used. (#490)
The text describes OAPH on line 99, but never defines what it means. It is left to the user to figure that out. In most cases that is fine, however, I feel it is more clear if explicitly mentioned on line 87 before it is used. Especially helpful to people new to all this jargon.